wip: update MemberTable
